# 结构设计规范知识库/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SourceBuildRequest": {
        "additionalProperties": false,
        "properties": {
          "apply_corrections": {
            "default": true,
            "title": "Apply Corrections",
            "type": "boolean"
          },
          "mode": {
            "default": "incremental",
            "pattern": "^(incremental|full)$",
            "title": "Mode",
            "type": "string"
          },
          "parser_backend": {
            "default": "mineru",
            "title": "Parser Backend",
            "type": "string"
          }
        },
        "title": "SourceBuildRequest",
        "type": "object"
      },
      "SourceBuildResponse": {
        "additionalProperties": false,
        "properties": {
          "job": {
            "additionalProperties": true,
            "title": "Job",
            "type": "object"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          }
        },
        "required": [
          "revision_id",
          "job"
        ],
        "title": "SourceBuildResponse",
        "type": "object"
      },
...
      "SourceDeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "source_id"
        ],
        "title": "SourceDeleteResponse",
        "type": "object"
      },
...
      "SourceMutationResponse": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "source"
        ],
        "title": "SourceMutationResponse",
        "type": "object"
      },
      "SourceResponse": {
        "additionalProperties": false,
        "properties": {
          "active_asset_version_id": {
            "default": "",
            "title": "Active Asset Version Id",
            "type": "string"
          },
          "active_governance": {
            "additionalProperties": true,
            "title": "Active Governance",
            "type": "object"
          },
          "active_metadata": {
            "additionalProperties": true,
            "title": "Active Metadata",
            "type": "object"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "governance": {
            "additionalProperties": true,
            "title": "Governance",
            "type": "object"
          },
          "lifecycle_status": {
            "title": "Lifecycle Status",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "pending_action": {
            "default": "",
            "title": "Pending Action",
            "type": "string"
          },
          "pending_asset_version_id": {
            "default": "",
            "title": "Pending Asset Version Id",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "updated_at": {
            "default": "",
            "title": "Updated At",
            "type": "string"
          },
          "versions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "source_id",
          "lifecycle_status",
          "metadata",
          "governance",
          "versions"
        ],
        "title": "SourceResponse",
        "type": "object"
      },
...
      "SourceUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "governance": {
            "additionalProperties": true,
            "title": "Governance",
            "type": "object"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          }
        },
        "title": "SourceUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "结构设计规范知识库 RAG API (多模态)",
    "version": "3.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/sources/candidates/{job_id}/republish": {
      "post": {
        "description": "Re-publish a failed candidate after re-running its activation gate only.",
        "operationId": "republish_source_candidate_admin_sources_candidates__job_id__republish_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceBuildResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Republish Source Candidate",
        "tags": [
          "source-management"
        ]
      }
    },
    "/admin/sources/changes/build": {
      "post": {
        "operationId": "build_source_changes_admin_sources_changes_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceBuildResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Source Changes",
        "tags": [
          "source-management"
        ]
      }
    },
...
    "/admin/sources/uploads": {
      "post": {
        "operationId": "upload_source_admin_sources_uploads_post",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 5,
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/pdf": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Source",
        "tags": [
          "source-management"
        ]
      }
    },
    "/admin/sources/{source_id}": {
      "get": {
        "operationId": "get_source_admin_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source",
        "tags": [
          "source-management"
        ]
      },
      "patch": {
        "operationId": "update_source_admin_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Source",
        "tags": [
          "source-management"
        ]
      }
    },
    "/admin/sources/{source_id}/discard-pending": {
      "post": {
        "operationId": "discard_pending_source_admin_sources__source_id__discard_pending_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Pending Source",
        "tags": [
          "source-management"
        ]
      }
    },
    "/admin/sources/{source_id}/draft": {
      "delete": {
        "operationId": "delete_draft_source_admin_sources__source_id__draft_delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Draft Source",
        "tags": [
          "source-management"
        ]
      }
    },
    "/admin/sources/{source_id}/retire": {
      "post": {
        "operationId": "retire_source_admin_sources__source_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire Source",
        "tags": [
          "source-management"
        ]
      }
    },
    "/admin/sources/{source_id}/validate": {
      "post": {
        "operationId": "validate_source_admin_sources__source_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Source",
        "tags": [
          "source-management"
        ]
      }
    },
    "/admin/sources/{source_id}/versions": {
      "post": {
        "operationId": "replace_source_admin_sources__source_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 5,
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/pdf": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Source",
        "tags": [
          "source-management"
        ]
      }
    },
...
  }
}